fix: Cleanup tests
