+++ src/internal/ui/zoxide/model_test.go
@@ -6,16 +6,46 @@ import (
 	tea "github.com/charmbracelet/bubbletea"
 	zoxidelib "github.com/lazysegtree/go-zoxide"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	"github.com/yorukot/superfile/src/internal/common"
 	"github.com/yorukot/superfile/src/internal/utils"
 )
 
+func TestHandleConfirmWithValidSelection(t *testing.T) {
+	m := setupTestModelWithResults(3)
+	m.cursor = 1
+
+	action := m.handleConfirm()
+
+	cdAction, ok := action.(common.CDCurrentPanelAction)
+	require.True(t, ok, "action should be CDCurrentPanelAction")
+	assert.Equal(t, m.results[1].Path, cdAction.Location, "action should navigate to results[1].Path")
+}
+
+func TestHandleConfirmWithNoResults(t *testing.T) {
+	m := setupTestModel()
+
+	action := m.handleConfirm()
+
+	_, ok := action.(common.NoAction)
+	assert.True(t, ok, "action should be NoAction when there are no results")
+}
+
+func TestHandleConfirmWithInvalidCursor(t *testing.T) {
+	m := setupTestModelWithResults(3)
+	m.cursor = 5
+
+	action := m.handleConfirm()
+
+	_, ok := action.(common.NoAction)
+	assert.True(t, ok, "action should be NoAction when cursor is out of bounds")
+}
+
 func TestJKKeyHandling(t *testing.T) {
 	m := setupTestModelWithClient(t)
 	common.Config.ZoxideSupport = true
 	m.Open()
-	m.justOpened = false
 
 	originalHotkeys := common.Hotkeys.ListDown
 	common.Hotkeys.ListDown = []string{"j", "down"}
